+++ efficientdet.egg-info/PKG-INFO
@@ -1,0 +1,4 @@
+Metadata-Version: 2.1
+Name: efficientdet
+Version: 0.1.0
+Summary: tfLITE
