+++ efficientdet.egg-info/SOURCES.txt
@@ -1,0 +1,19 @@
+MANIFEST.in
+README.md
+requirements.txt
+setup.py
+efficientdet/__init__.py
+efficientdet/app.py
+efficientdet/demo.py
+efficientdet/dog.jpg
+efficientdet.egg-info/PKG-INFO
+efficientdet.egg-info/SOURCES.txt
+efficientdet.egg-info/dependency_links.txt
+efficientdet.egg-info/entry_points.txt
+efficientdet.egg-info/requires.txt
+efficientdet.egg-info/top_level.txt
+efficientdet/models/__init__.py
+efficientdet/models/model.py
+efficientdet/models/model_float32.5
+tests/__init__.py
+tests/regression_test.py
