+++ efficientdet.egg-info/entry_points.txt
@@ -1,0 +1,2 @@
+[console_scripts]
+efficientdet = efficientdet.demo:main
