+++ efficientdet.egg-info/requires.txt
@@ -1,0 +1,45 @@
+absl-py==1.4.0
+astunparse==1.6.3
+cachetools==5.3.0
+certifi==2022.12.7
+charset-normalizer==3.0.1
+flatbuffers==23.1.21
+gast==0.4.0
+google-auth==2.16.0
+google-auth-oauthlib==0.4.6
+google-pasta==0.2.0
+grpcio==1.51.1
+h5py==3.8.0
+idna==3.4
+importlib-metadata==6.0.0
+keras==2.11.0
+libclang==15.0.6.1
+Markdown==3.4.1
+MarkupSafe==2.1.2
+numpy==1.24.2
+oauthlib==3.2.2
+opt-einsum==3.3.0
+packaging==23.0
+Pillow==9.4.0
+protobuf==3.19.6
+pyasn1==0.4.8
+pyasn1-modules==0.2.8
+requests==2.28.2
+requests-oauthlib==1.3.1
+rsa==4.9
+setuptools>=67.3.2
+six==1.16.0
+streamlit
+tensorboard==2.11.2
+tensorboard-data-server==0.6.1
+tensorboard-plugin-wit==1.8.1
+tensorflow==2.11.0
+tensorflow-estimator==2.11.0
+tensorflow-io-gcs-filesystem==0.30.0
+termcolor==2.2.0
+tflite==2.10.0
+typing-extensions==4.4.0
+urllib3==1.26.14
+Werkzeug==2.2.2
+wrapt==1.14.1
+zipp==3.13.0
