+++ efficientdet.egg-info/top_level.txt
@@ -1,0 +1,2 @@
+efficientdet
+tests
