+++ setup.py
@@ -6,15 +6,15 @@ This script installs the demo package and its dependencies.
 from setuptools import find_packages, setup
 
 setup(
-    name="demo",
+    name="efficientdet",
     version="0.1.0",
     packages=find_packages(),
     entry_points={
         "console_scripts": [
-            "demo=demo.demo:main",
+            "efficientdet=efficientdet.demo:main",
         ]
     },
-    package_data={"demo": ["models/*", "dog.jpg"]},
+    package_data={"efficientdet": ["models/*", "dog.jpg"]},
     description="tfLITE",
     include_package_data=True,
     install_requires=open("requirements.txt", "r", encoding="utf-8").readlines(),
